refactor: rename hooks.ts to clientHooks.ts
